Fixed a segmentation fault that was occurring on logout
This issue is very similar to what was fixed in 66de91f. The
segmentation fault was occurring upon logout when a user was logged in,
and a logout request was sent without nullifying the callbacks of all
other HTTP connections. This led to the main event structure continuing
to executed with invalid memory pointers.