steam: better handling of SteamGuard authentication
When a SteamGuard authentication code is required, logout of the
account to prevert the BitlBee timeout from logging it out. Then wait
for the user to enter their code with the account offline, and upon
entry of the code, turn the account on and proceed.

This addresses a timing out issue as well a segfault that would occur
with the old SteamGuard authentication handling.