steam-api: removed local friend storage, store them within BitlBee
Storing steams friends was previously done locally in the SteamAPI,
which only does good in one region. Storing the friends with BitlBee
prevents doubly storing friends, and it also allows offline friends
to be displayed. Friend storage is also needed to ensure consistency
of the buddy list (see: 6836d26).  Furthermore, it is a lot cleaner.

This resolves offline Steam Friends not being displayed on demand.