steam: added 'show_playing' set variable
This allows for game play status updates to be toggled off or switched
from half-op mode.

Several changes had to be made to do this correctly, the biggest of
which was dropping the PLAYING Steam state. In doing so, an issues with
away states was resolved, and allows for a friend to playing and away
at the same time.