steam: removed dynamic polling, use keep-alive sessions instead
The plugin already uses the keep-alive method for polling as it is. The
dynamic polling was just getting in the way by doing absolutely nothing
worthwhile. Furthermore, the timeout value should have been in
milliseconds, not seconds.

The keep-alive HTTP requests are kept open until a message is sent, and
it allows for an "instant" retrieval of messages. In addition, it means
less HTTP requests are need. After 30 seconds, the maximum timeout, a
timeout "error" is returned and handled gracefully, then a new HTTP
request is initiated.

However, this does seem pose an issue after long uptimes. The internal
HTTP client seems to hangup on these keep-alive sessions, and when the
Steam account is disabled, it crashes. But, it cannot hurt to have this
method implemented without the old bloat.