steam-http: free SteamHttpReq data unless specified otherwise
This simply adds a field to SteamHttpReq that is called upon freeing the
request in order to ensure the data is freed. A default value for this
field is stored in SteamHttp and is set upon initialization.

This fixes a minor memory leak that would occur when connections were
forcibly closed with steam_http_req_free() or steam_http_free_reqs().