steam: update nicknames and real names locally when they are changed
When a `personastate` poll event is received, a check is now performed
to make sure the buddy's nickname, and optionally the buddy's real name,
have not changed. If change has occurred, updates are handled locally.

A set variable known as `extra_info` has been added to toggle the
summary requests that must be sent for full names and game statues.

This implements functionality of storing the real names of buddies.