steam: handle `personarelationship` poll events
The first sign of managing friendships via this plugin. The API event
allows for the "synchronization" of the friendship statuses with the
Steam Community. The implementation for this is rather sloppy as the
API does not return the current nick for the friend in the event.

This does not actually enable the ability to manage friendships via
this plugin, however, it does enable status updates/notifications.