steam-api: fixed poll reply messages being reversed
The issue was introduced with 259c2bd. A simple fix to the issue is to
reverse the GSList that holds the SteamMessages.

A "better" but less efficient method would be to sort the GSList by the
timestamp of each SteamMessage. This would ensure order consistency
without relying on the poll reply's default order. This will be
implemented if more related issues to message ordering arise.

This fixes messages being sent in a reversed order.