steam-api: fixed the last message id not being parsed and sent
The last message id, which is needed by poll requests in order to
request new messages, was not being grabbed correctly from the JSON
parser. The JSON parser does not doubly story an integer's value in
string form. This means the plugin needs to handle the actual integer
value for the JSON object, rather than attempting to use strings.

This resolves issues related to messages not being received from polling
operations.