steam-api: decrease the poll timeout to 15 seconds
A poll timeout of 30 seconds appears to be to long, and allows for the
API session to expire very easily. With a timeout of 15 seconds, the API
session appears to keep valid until an actual timeout barrier is hit.

This will help or completely alleviate poll messages not being received.