steam: reauthenticate after the account password is changed
After a Steam account's password is changed, all API requests return an
HTTP 401 (unauthorized) error. This is due to the fact that the API
access token no longer valid.

This issue is resolved by adding a set evaluator for the password set.
This will allows the plugin to know when the user's password has been
changed. Upon the password being changed, the current API access token
is discarded, the account is reconnect, and then the reauthentication
process beings.