steam: removed the 'extra_info' set
This set was originally introduced to make the secondary user summary
request togglable. This option defaults to true, which means *most*
users will never turn this setting off. Additionally, this setting is
needed for the game status updates of Steam Friends, which is one of
the major points to using Steam's chat interface.

There is not much practically in keeping an option around that bloats
the code by ~50 lines if it is likely to never be used.