steam: reduced user mode spam upon changing game statuses
Currently, it is assumed that a new personastate event means a friend's
state actually changed, however, this is not always the case. A
personastate event is dispatched if a friend joins or changes gamer
servers, and this information is not relevant during poll operations.
Additionally, a large sum of repeated personastate events are sent when
the actual Steam client connects.

Since a flag cannot be passed to imcb_buddy_status() like it can be with
away states, everything is handled within the plugin. This means the
plugin needs to ensure that the game status is not already defined.