steam-http: handle NULL request pointers from http_dorequest()
There is a chance of http_dorequest() returning a NULL pointer for the
request. Currently, this is not accounted for, and results in the
request going missing without a trace. Rather than just ignoring the
failed request, call the callback provided with an error.

This resolves some requests going missing without a trace, and never
getting processed.