steam-api: implemented the new Steam Mobile authentication API
The Steam Mobile API has recently changed, and the older authentication
API has been removed. The main purpose of the new authentication API
appears to be security enhancements. This patch supports all of the
known security enhancements.

Security enhancements within the new API:
  1. Captcha requirement after too many failed logins
  2. Encryption of the account password

A new dependency, GMP, has been introduced and is used for a local
implementation of PKCS1. This prevents the plugin from having to depend
on other encryption libraries, which BitlBee's configuration headers do
not share.

This resolves issues with account authentication failing due to an HTTP
404 error.