steam-api: assign all callback functions within the SteamApi
Currently the callback functions are assigned when a SteamApi function
is called. Rather than defining the callbacks every time, define them
when the SteamApi is initialized. This is cleaner as well as allows
for greater functionality. With the old method, there was no clean or
consistent way of defining a callbacks that did not pertain to an API
function.