steam: do not reset the UMQID on login
Currently, a logout request is sent for the UMQID before the login
request is sent. This was done to reset the UMQID, and ensure the UMQID
was logged out before attempting to login. This operation is no longer
needed as the login request appears to now override any older sessions.