steam-http: fixed requests randomly hanging
With the current HTTP interface, there is no timeout for faulty
connections. This is especially problematic with the use of long
standing keep-alive requests that go bad.

This patch uses a timed event to watch after each request. While it
would be more efficient to hook right into glib/libevent and handle
timeouts, it is not as portable.