steam-http: refactored the internal storage of headers and parameters
Currently, the plugin stores escaped values within the internal storage
structures. This is problematic when a value needs to be retrieved or
modified. Storing unescaped values will help keep the HTTP interface
sane and consistent when it is extended in the future.

Additionally, many months ago with e2fb61f the plugin was switched from
GTrees to GHashTables. This seemed like a good idea at the time, but
with all of the traversing, GTrees are in reality a better option.