steam: fixed CTCP spam from incoming messages
With each new message, the typing state is set to a false state. This
dispatches a CTCP message each time leading to a great deal of CTCP
spam. As a result, the typing state is to only be set to a false state
if the typing state is currently active.