Retrieve unread messages upon connecting
With the introduction of the Steam Community Web Chat, the chat log
can now be retrieved via the Steam Web API. This allows the plugin to
retrieve unread messages and mark them as read.

With the Steam Web API, the amount of unread messages and a timestamp
for the last read message are provided. However, the plugin makes use
of the OAuth API, rather than the web API. This means the plugin has to
check each user for unread messages upon connecting which leads to a lot
of initial HTTP requests.

In addition, a timestamp has to be stored locally for the last read
message, rather using the true last read message timestamp. If multiple
Steam chat clients are being used, this can prove problematic with some
messages still being seen as unread when in reality they are not.