steam-api: use the 'web' chat mode
By default when logging on via this plugin, the user is shown as being
online via mobile device. With the introduction of the Steam Community
Web Chat, there is now a 'web' chat mode.

Since the 'mobile' chat mode is rather misleading, this plugin will now
make use of the 'web' chat mode.