Fixed relogon operations not resending failed requests
When a relogon request is sent and processes successfully, the requests
that failed are not resent correctly. This is noticeable with polling
requests as they are never reinitiated. As a result, the polling loop
comes to an end which leads to no messages being retrieved. Moreover,
this kills the logon session, and eventually the user will appear
offline. Additionally, queued messages fail to resend correctly.