steam: do not handle timestamps on polled messages
Originally, when timestamp handling was implemented, it was done with
the intent of knowing exactly when that message was created. However,
if anything, the handling timestamps on polled messages has just led to
problems. Timestamps are not relevant when holding a conversation, even
when latency gets in the way. Additionally, no other BitlBee protocol
plugin handles timestamps on polled messages, so this plugin shall do
the same.

This does *not* remove the handling of timestamps on the unread messages
retrieved upon login, only polled messages.