steam: immediately store the last message timestamp
Currently, the last message timestamp is stored locally, and only stored
in the set interface upon the account logging off. This becomes an issue
when BitlBee is unexpectedly killed, and the account was not logged off.

The solution is to directly work with the set interface for storing the
last message timestamp. It is possible this will not completely address
the issue, and a force save will need to be done when the set value is
updated. However, this is not implemented in this patch, and will only
be done as as last resort due to the overhead.