steam-api: error upon low poll timeouts
After a period of time, poll requests will begin to timeout after one
second, rather than the set timeout value. This leads to a great deal
of HTTP spam which is completely unwarranted. While the exact cause is
unknown, it is probably related to the relogon when the current Steam
API session expires.

A check for this error must be implemented to prevent excessive HTTP
requests. Moreover, even after the root cause has been determined, and
the issue has been resolved, it should be ensured the timeout is not
too low.