steam: Fixed setting `show_playing` during account runtime
The plugin is suppose to revalidate the buddy list of an account that is
online when `show_playing` is changed. Currently, the plugin neglects to
do this correctly, and the buddy list is not revalidated at all.

Additionally, any invalid user mode should disable the feature. While
this worked, a check was added to ensure an invalid mode was not set.