steam-api: retrieve extra timestamps for friends
The steam_api_friends() function will now populate friend summaries with
the last message and the last view timestamps. This is done by parsing
the values from the Steam Web Chat page. This is certainly not the best
method, but it is the only method currently known. The new timestamp
values make it possible for the plugin to determine if there are unread
messages, and if so, which messages are unread.

This removes the hackery of locally storing the timestamp of the last
message that was polled. Moreover, this ensures that only truly unread
messages are retrieved as well as significantly reducing HTTP requests.