steam-api: fixed the low poll timeout error check
Currently, if no messages are retrieved from a poll response, a check
is initiated on the timeout to ensure it is not too low. However, a
response with an empty message list can break this check. The simple
solution is to only check the poll timeout when the response has a
timeout error.

This fixes the sending of a message triggering the low poll timeout
error message.