steam-api: error upon empty summary results
Often times, summary requests are needed to retrieve additional friend
information. However, the Steam API occasionally returns an empty list
of summaries. This leads to a slew of problems for the plugin.

This patch fixes several issues:
  - Endless amount of summary requests
  - Huge amounts of HTTP request spam
  - The halting of certain API requests

The solution is to simply error when this issue occurs, and reconnect,
which typically resolves the issue.