steam-glib: fixed compatibility implementations
As of glib-2.39.0, g_hash_table_add() now returns a boolean. However,
this is not the issue, this just happened to uncover a rather serious
issue. The way the plugin was overriding missing glib symbols was
incorrect, leading to things being double defined in some situations.

The solution is simple, use the versioning macro in glib, rather than
checking the existence of each symbol.