Fixed relogon and the underlying request resending
Currently, requests that are queued for resending when the queue is
paused are immediately sent as well as requeued. This is problematic,
as they are resent without the STEAM_HTTP_REQ_FLAG_NOFREE flag, and
this inevitably leads to segmentation faults. Moreover, the whole
resending logic needed a great deal of refactoring.

Additionally, much refactoring was done on the steam-api interface to
conform with similar hidden operations. This keeps the steam-api
interface consistent.

This fixes the client crashing upon the relogon operation.