steam-friend: fixed channel modes not being overridden
When steam_friend_chans_umode() is used to set the channel mode(s) of a
SteamFriend, the mode(s) do not override existing modes. This prevents
modes of a lower standing from being set correctly.

This patch will simply unset all other mode flags that are a higher
standing than that of the mode(s) being set.