steam: fixed segfault when setting `show_playing`
This segfault occurs from the account or bee_user being from a different
account or invalid. This issue seems to only be noticeable while the
steam account and a skype account are both online.

In order to resolve this issue, conditional checks have been implemented
to ensure the account and bee_user are valid.