steam-friend: properly parse and store SteamIDs/CommunityIDs
This significantly cleans up and unifies the SteamID/CommunityID parser
and storage methods. Previously, IDs were incorrectly parsed and there
was a lot of information missing, such as the ID type and universe. This
will enable the plugin to fix current issues and implement features down
road where this information is essential. In addition, this prevents the
plugin from having to constantly reparse IDs all over the place as IDs
are now centralized in one object.