steam-api: denote non-Steam games via the game name
The Steam Web API will send the game ID along with the game name if and
only if the game is a Steam game. Otherwise, only the game name will be
sent. This makes it rather trivial to detect if a game is a Steam game
or not. If a game is not a Steam game, it should denoted as such, much
like in the main Steam client.