steam-api: fixed authentication failures with captcha and SteamGuard
There is a global error check for all API requests and their replies,
which was preventing authentication information from being retrieved
and stored. This caused the SteamApi interface to fail when captcha or
SteamGuard information was required. The fix for this is simple, just
ignore these errors on a global level, and handle them locally in the
authentication callback.