steam-http: fixed case-insensitive comparisons failing for GHashTables
Currently, the GEqualFunc given to GHashTables is g_ascii_strcasecmp(),
however, this does not match the required function signature. The return
value should be a gboolean, rather than a comparison integer.

This addresses a slew of issues, namely values not being set and updated
correctly. This led to issues such as the account requiring a fresh
session identifier at each login since it was not stored correctly.