steam-api: replace XML character entities with certain requests
Some of the Steam Community API requests provide XML encoded character
entities for displaying the data directly on web pages. This does not
apply to this plugin since nothing is render via HTML. As a result, the
XML entities must be replaced with their representing character to have
seamless integration.