steam-crypt: implemented cryptographic interface
Currently, the plugin handles some cryptographic operations internally
via GMP. However, as time has progressed, it is has become evident that
this is not practical. This was done initially to prevent from having to
implement and maintain the three different SSL libraries which bitlbee
supports.

Advantages of libgcrypt:
  - Available on most systems
  - Has the required cryptographic implementations
  - Prevents the implementation of three SSL libraries
  - Required by bitlbee for OTR support

The cryptographic interface is based on libgcrypt. As a result of
introducing libgcrypt, GMP is no longer needed.