steam-id: implemented the SteamID interface
This implements an elegant interface for working with 64-bit SteamIDs.
The interface is implemented with only a simple series of macros, which
eliminate the need of storing the SteamID in abstract data formats. As
are result, the steam-id interface requires very little memory.

This removes the existing SteamID implementation within the steam-user
interface.