steam: fixed potential NULL pointer dereference with user statuses
There was a rare corner cases that might allow for the dereference of a
NULL pointer with steam_user_status(). The steam_user_status() function
should gracefully attempt to grab the bee_user_t if it is NULL, rather
than relying on the caller to NULL check the bee_user_t.