configure: bump required bitlbee version to 3.2.2
As of bitlbee-3.2.2, the pkg-config file now contains properly defined
variables. The plugin can now obtain the plugin directory and check the
version via pkg-config. In addition, the version bump allows the plugin
to remove some compatibility code in the steam-json interface.

This fixes a compile time warning in the steam-json interface.