steam-crypt: handle erroneous libgcrypt calls
Currently, errors produced by many of the libgcrypt calls are simply
ignored. This can lead to segmentation faults within libgcrypt as the
plugin will continue to operate with libgcrypt and make other library
calls with invalid pointers. All libgcrypt calls should be checked for
errors and/or pointers which are invalid.

This addresses the issue of segmentation faults occurring with older
libgcrypt versions during the authentication phase.