steam-api: use strip_html() for unescaping XML entities
As of now, the plugin uses an internal function for unescaping the XML
entities from a string. However, bitlbee offers a public function for
stripping HTML from a string, which also unescapes the entities. Rather
than using an internal function, use the bitlbee provided function in
order to reduce code.

Thanks to @seirl for pointing this out.