Implemented the setting of the user state
It was recently discovered that the state of a user can be defined by
setting the idle timeout via poll requests. If the timeout exceeds five
minutes in seconds, the user will be flagged as away. Additionally, if
the timeout exceeds two hours in seconds, the user will be flagged as
snoozing. While this will mark the user as being away or snoozing, it
is limited to only these two states. The user is not able to set any
other Steam state, such as busy.

The plugin makes use of this timeout by forcibly setting the timeout to
a static value well over the required value for good measure.

It should be noted that the plugin does cancel any outstanding polling
operations in order to apply the new state. The plugin will wait for
the outstanding polling operation to complete. While this prevents any
sort of overly complex poll tracking, it also means it can take up to
30 seconds before the state is updated.