Migrated to the newer chat history API
This migrates the chat history implementation from using the Steam
Community API to using the newer Mobile API. This is not only cleaner,
but it is far more reliable.

Marking unread messages as read is broken with this newer API, which
leads to unread messages being repeatedly displayed. The request that
is used to mark messages as read still needs to be reversed. This will
be done at a later time, as this is an urgent release to get the plugin
back online. In the mean time, users can simply go to the website in
order to mark the messages as read.

This resolves certain requests to the Steam Community failing, as the
content is now forcibly being gzipped.