steam-api: fixed OAuth data failing to parse
It appears the mobile login request can return a completely JSONized
object, rather than JSON in a JSON string. The problem with this is
that, it does not allow the plugin to use the token for OAuth style
requests.

In order to continue getting the JSON string, additional cookies are
needed to identify the device as mobile.

This fixes authentications requests failing due to the JSON string
being unparsable JSON.