Reimplemented marking message as read
Marking messages as read regressed as a result of a previous API update
(485ee3d). It was initially thought that a newer API would replace the
usage of the Steam Community API. However, this newer API appears to be
nothing more than a dead-end. To keep this functionality operational
until the newer API is reversed, usage of the Steam Community API has
been reintroduced.